refactor: renamed auth context files for clarity
