+++ client/src/components/ProtectedRoute.jsx
@@ -1,5 +1,5 @@
 import { Navigate, useLocation } from "react-router-dom";
-import { useAuth } from "../context/auth-context";
+import { useAuth } from "../context/useAuth";
 
 export default function ProtectedRoute({ children }) {
   const { isAuthenticated, isLoading } = useAuth();
